Split endpoint engine tests module
